Remodel projects into library & samples
