+++ EntryAutoComplete/EntryAutoComplete.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EntryAutoComplete", "EntryA
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EntryAutoComplete.UWP", "EntryAutoComplete\EntryAutoComplete.UWP\EntryAutoComplete.UWP.csproj", "{A8CDB6C9-6DB4-4B4D-BADC-8FF7CE0A3DAD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EntryAutoComplete.Sample", "EntryAutoComplete.Sample\EntryAutoComplete.Sample.csproj", "{71A1C6B8-C399-4AEB-95B3-40C2638F4854}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Ad-Hoc|Any CPU = Ad-Hoc|Any CPU
@@ -279,6 +281,54 @@ Global
 		{A8CDB6C9-6DB4-4B4D-BADC-8FF7CE0A3DAD}.Release|x86.ActiveCfg = Release|x86
 		{A8CDB6C9-6DB4-4B4D-BADC-8FF7CE0A3DAD}.Release|x86.Build.0 = Release|x86
 		{A8CDB6C9-6DB4-4B4D-BADC-8FF7CE0A3DAD}.Release|x86.Deploy.0 = Release|x86
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|Any CPU.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|Any CPU.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|ARM.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|iPhone.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|iPhone.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|x64.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|x64.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|x86.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Ad-Hoc|x86.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|Any CPU.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|Any CPU.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|ARM.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|ARM.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|iPhone.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|iPhone.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|x64.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|x64.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|x86.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.AppStore|x86.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|ARM.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|x64.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Debug|x86.Build.0 = Debug|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|Any CPU.Build.0 = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|ARM.ActiveCfg = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|ARM.Build.0 = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|iPhone.Build.0 = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|x64.ActiveCfg = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|x64.Build.0 = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|x86.ActiveCfg = Release|Any CPU
+		{71A1C6B8-C399-4AEB-95B3-40C2638F4854}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
