+++ EntryAutoComplete/EntryAutoComplete/EntryAutoComplete.Android/MainActivity.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-
 using Android.App;
 using Android.Content.PM;
 using Android.Runtime;
@@ -19,7 +18,7 @@ namespace EntryAutoComplete.Droid
 
             base.OnCreate(savedInstanceState);
             global::Xamarin.Forms.Forms.Init(this, savedInstanceState);
-            LoadApplication(new App());
+            LoadApplication(new global::EntryAutoComplete.Sample.App());
         }
     }
 }
