+++ EntryAutoComplete/EntryAutoComplete/EntryAutoComplete.UWP/EntryAutoComplete.UWP.csproj
@@ -104,7 +104,6 @@
     <AppxManifest Include="Package.appxmanifest">
       <SubType>Designer</SubType>
     </AppxManifest>
-    <None Include="EntryAutoComplete.UWP_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="baseline_close_black_24.png">
@@ -138,9 +137,9 @@
     </PackageReference>
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\EntryAutoComplete\EntryAutoComplete.csproj">
-      <Project>{d53a14f2-2ebd-45ab-8e44-517b1ee6679b}</Project>
-      <Name>EntryAutoComplete</Name>
+    <ProjectReference Include="..\..\EntryAutoComplete.Sample\EntryAutoComplete.Sample.csproj">
+      <Project>{71a1c6b8-c399-4aeb-95b3-40c2638f4854}</Project>
+      <Name>EntryAutoComplete.Sample</Name>
     </ProjectReference>
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
