+++ EntryAutoComplete/EntryAutoComplete/EntryAutoComplete.iOS/AppDelegate.cs
@@ -23,7 +23,7 @@ namespace EntryAutoComplete.iOS
         public override bool FinishedLaunching(UIApplication app, NSDictionary options)
         {
             global::Xamarin.Forms.Forms.Init();
-            LoadApplication(new App());
+            LoadApplication(new global::EntryAutoComplete.Sample.App());
 
             return base.FinishedLaunching(app, options);
         }
