+++ EntryAutoComplete/EntryAutoComplete/EntryAutoComplete.iOS/Renderers/ShadowedFrameRenderer.cs
@@ -1,5 +1,5 @@
 ﻿using CoreGraphics;
-using EntryAutoComplete.CustomControl;
+using EntryAutoComplete;
 using EntryAutoComplete.iOS.Renderers;
 using UIKit;
 using Xamarin.Forms;
